Add error handling to catch if file passed in does not exist
On event that the file passed in to the program does not exist, the
program will catch the error and print out a simple message notifying
the user that the file does not exist.

Updated usage printout to specify python3 should be used rather than
python when calling program.  If python3 is not used, the error handling
added in this commit will not take effect.